Unify service worker logic
