+++ apps/extension/src/service-workers/background.ts
@@ -1,110 +1,142 @@
-import { CreateNavigationEntry } from '@wtm/api';
-import { DOMtoString, getImages, isTokenExpired } from '@wtm/utils';
+import { CreateMLCEngine, MLCEngineInterface } from '@mlc-ai/web-llm';
+import { isTokenExpired } from '@wtm/utils';
 import { apiClient } from '../utils/api.client';
-import * as cheerio from 'cheerio';
-import { convertHtmlToMarkdown } from 'dom-to-semantic-markdown';
-import { AnyNode } from 'domhandler';
 
-const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
+import {
+  EngineStatus,
+  Ports,
+  ServiceWorkerMessageType,
+  ServiceWorkerPayload,
+} from './types';
 
-const getSemanticMarkdownForLLM = (
-  data: string | AnyNode | AnyNode[] | Buffer,
-) => {
-  const $ = cheerio.load(data);
-  $('script, style, nav, footer, header, .ads, .banner').remove();
+let engine: MLCEngineInterface | undefined = undefined;
+let engineStatus: EngineStatus = EngineStatus.NOT_READY;
 
-  $('code').each((_, elem) => {
-    const codeText = $(elem).html();
-    $(elem).html(`<code>${codeText}</code>`);
-  });
-
-  $('body *').each((_, elem) => {
-    if ($(elem).is('button') || $(elem).is('div[onclick]')) {
-      $(elem).remove();
-      return;
-    }
-    const text = $(elem).text().trim();
-    if (!text || text.length < 5) {
-      $(elem).remove();
-    }
-
-    // Unify links test in one line
-    if ($(elem).is('a')) {
-      const linkText = $(elem).text().replace(/\s+/g, ' ').trim(); // Replaces multiple spaces for an empty space.
-      $(elem).text(linkText); // Updates text content from an <a>
-    }
-  });
-  const content = $('body').html();
-  return convertHtmlToMarkdown(content || '');
+let intervalId: NodeJS.Timeout;
+const startInterval = () => {
+  intervalId = setInterval(refreshAccessToken, 60 * 60 * 1000);
 };
 
-chrome.tabs.onUpdated.addListener(async (tabId, changeInfo, tab) => {
-  try {
-    const { accessToken, enabledLiteMode, stopTrackingEnabled } =
-      await chrome.storage.local.get([
-        'accessToken',
-        'enabledLiteMode',
-        'stopTrackingEnabled',
-      ]);
+/**
+ * Adds a listener for incoming connections to the service worker.
+ * @param {chrome.runtime.Port} port - The port connecting to the service worker.
+ */
+chrome.runtime.onConnect.addListener(async (port) => {
+  if (port.name !== Ports.SERVICE_WORKER) {
+    console.log('Invalid port name');
+    return;
+  }
 
-    if (stopTrackingEnabled) return;
+  const { webLLMEnabled } = await chrome.storage.local.get(['webLLMEnabled']);
 
-    if (accessToken && changeInfo.status === 'complete' && tabId) {
-      if (tab.url && !tab.url.startsWith('chrome://')) {
-        await sleep(300);
-        tab = await chrome.tabs.get(tabId);
+  if (engineStatus === EngineStatus.NOT_READY && webLLMEnabled) {
+    try {
+      await initEngine();
+      port.postMessage({
+        type: ServiceWorkerMessageType.ENGINE_READY,
+      });
+    } catch (error) {
+      console.log(error);
+    }
+  }
 
-        const images = await chrome.scripting.executeScript({
-          target: { tabId: tabId },
-          func: getImages,
+  /**
+   * Handles messages received on the port.
+   * @param {ServiceWorkerPayload} message - The message received from the port.
+   */
+  port.onMessage.addListener(async (message: ServiceWorkerPayload) => {
+    switch (message.type) {
+      case ServiceWorkerMessageType.GENERATE_COMPLETION: {
+        if (!engine) {
+          console.error('Engine is not ready');
+          return;
+        }
+
+        const result = await engine.chat.completions.create({
+          messages: [
+            {
+              role: 'system',
+              content:
+                "You will be provided with a webpage's content in semantic markdown format. Your task is to analyze the markdown and generate a brief summary of the key points in 3 to 5 sentences. The summary should be concise and focus on the most important information presented in the content.",
+            },
+            {
+              role: 'user',
+              content: message.content,
+            },
+          ],
         });
 
-        let results;
-        if (!enabledLiteMode) {
-          try {
-            results = await chrome.scripting.executeScript({
-              target: { tabId: tabId },
-              func: DOMtoString,
-              args: ['body'],
-            });
-          } catch (error) {
-            return;
-          }
+        const response = result.choices.at(0)?.message.content;
 
-          // If body it's present in DOM, we take it from the inital position.
-          const htmlContent = results[0].result || '';
+        const url = message.url;
 
-          const content =
-            enabledLiteMode == undefined || !enabledLiteMode
-              ? getSemanticMarkdownForLLM(htmlContent)
-              : '';
-          const navigationEntry: CreateNavigationEntry = {
-            url: tab.url!,
-            navigationDate: new Date().toISOString(),
-            title: tab?.title || '',
-            // Removes remaining HTML tags, more than 2 contiguous spaces, more than 2 contiguous line breaks, and HTML entities
-            ...((enabledLiteMode === undefined || !enabledLiteMode) && {
-              ...(content && { content }),
-              images: images ? images[0].result! : [],
-            }),
-          };
+        await apiClient.securedFetch('/api/navigation-entry/add-context', {
+          method: 'POST',
+          body: JSON.stringify({
+            content: response,
+            url,
+          }),
+        });
 
-          await apiClient.securedFetch('/api/navigation-entry', {
-            method: 'POST',
-            body: JSON.stringify(navigationEntry),
-          });
-        }
+        break;
       }
+
+      case ServiceWorkerMessageType.CREATE_NAVIGATION_ENTRY: {
+        const navigationEntry = message.navigationEntry;
+
+        await apiClient.securedFetch('/api/navigation-entry', {
+          method: 'POST',
+          body: JSON.stringify(navigationEntry),
+        });
+
+        break;
+      }
+
+      case ServiceWorkerMessageType.UPDATE_EXTENSION_ICON:
+        setCorrectIconByUserPreferences();
+        break;
+
+      default:
+        break;
     }
-  } catch (error) {
-    console.error(`Unexpected Error in tabs onUpdated:`, error);
+  });
+});
+
+chrome.windows.onFocusChanged.addListener(async (windowId) => {
+  if (windowId !== chrome.windows.WINDOW_ID_NONE) {
+    clearInterval(intervalId);
+    startInterval();
   }
 });
 
-let intervalId: NodeJS.Timeout;
+chrome.runtime.onStartup.addListener(() => {
+  refreshAccessToken();
+  startInterval();
+});
 
-const startInterval = () => {
-  intervalId = setInterval(refreshAccessToken, 60 * 60 * 1000);
+chrome.runtime.onInstalled.addListener(() => {
+  refreshAccessToken();
+  startInterval();
+});
+
+/**
+ * Initializes the MLCEngine for use in generating completions.
+ * @returns {Promise<void>} A promise that resolves when the engine is initialized.
+ */
+const initEngine = async () => {
+  try {
+    engine = await CreateMLCEngine('SmolLM-360M-Instruct-q4f16_1-MLC', {
+      initProgressCallback: () => {
+        if (engineStatus === EngineStatus.NOT_READY) {
+          engineStatus = EngineStatus.LOADING;
+        }
+      },
+    });
+
+    engineStatus = EngineStatus.READY;
+  } catch (error) {
+    console.error('Error initializing engine', error);
+  }
 };
 
 const defaultIcons = {
@@ -122,12 +154,20 @@ const grayScaleIcons = {
 };
 
 const noTrackingIcons = {
-  '16': 'image.png',
-  '32': 'image.png',
-  '48': 'image.png',
-  '128': 'image.png',
+  '16': 'app-icon-no-tracking-16.png',
+  '32': 'app-icon-no-tracking-32.png',
+  '48': 'app-icon-no-tracking-48.png',
+  '128': 'app-icon-no-tracking-128.png',
 };
 
+/**
+ * Sets the extension icon based on the user's preferences.
+ * If the "stop tracking" preference is enabled, a no-tracking icon is used.
+ * Otherwise, a default icon is set.
+ * @async
+ * @function setCorrectIconByUserPreferences
+ * @returns {Promise<void>} Resolves when the icon has been set.
+ */
 const setCorrectIconByUserPreferences = async () => {
   const response = await apiClient.getUserPreferences();
   if (response?.enableStopTracking) {
@@ -141,6 +181,14 @@ const setCorrectIconByUserPreferences = async () => {
   }
 };
 
+/**
+ * Refreshes the access token if needed and updates the extension icon accordingly.
+ * If the access token is expired and the refresh token is still valid, it refreshes the access token.
+ * If no tokens are found or both are expired, the icon is set to grayscale.
+ * @async
+ * @function refreshAccessToken
+ * @returns {Promise<void>} Resolves when the token check and potential refresh are complete.
+ */
 const refreshAccessToken = async () => {
   try {
     const { accessToken, refreshToken } = await chrome.storage.local.get([
@@ -173,15 +221,3 @@ const refreshAccessToken = async () => {
     console.error(`Unexpected Error in windows onFocusChanged:`, error);
   }
 };
-
-chrome.windows.onFocusChanged.addListener(async (windowId) => {
-  if (windowId !== chrome.windows.WINDOW_ID_NONE) {
-    clearInterval(intervalId);
-    startInterval();
-  }
-});
-
-chrome.runtime.onStartup.addListener(() => {
-  refreshAccessToken();
-  startInterval();
-});
